refactor(settings): extract marking-types-settings into subcomponents
